feat: Add TinyGo to the toolchain with build-tool
The build-tool downloads a TinyGo release archive, checks for a
known-good SHA-256 value and extracts the release.

This change also renames build-tools to build-tool and
scripts/bootstrap-build-tools.sh to scripts/bootstrap-build-tool.sh.
The toolchain has moved from ./build-tools to ./toolchain.

The Makefile and README.md have been updated to reflect the use of
build-tool.