build: update tinygo's modification time
Make looks at a file's modification time to determine whether a target
is up-to-date.  Because build-tool had been restoring file modification
times after extracting archives, make did not see the TinyGo target as
up-to-date.  This works around that by changing the modification time of
the TinyGo executable.  (An alternative would be to create a sentinel
file at the end of the process, which would serve as the make target.)