fix: build Cap'n Proto despite the capnp directory
The default executable for Cap'n Proto is "capnp".  This is also the
name of a directory in the root of the project.  `fileExistsAtPath()`
was finding that "capnp" (the directory) existed instead of noting that
the Cap'n Proto compiler "capnp" did not exist, thus it was refusing to
build the Cap'n Proto compiler.  This change looks only for files of
this name.  I'm not certain that this is the best fix, but it solves the
problem at the moment.