refactor: better experiences for failures
fix: remove empty file when download fails
The empty file was causing a future run of the script to fail.

fix: add error code for failed download
We have many error codes but we missed this one somehow.

refactor: add max_delay to retries
If retryable_curl is going to ask you to retry every time, we know that
it will not be DOSing the download servers.

fix: report the path when rejecting the SHA-256
Before this change, it only reported the file name, which is not in the
working directory.